requirements.txt
#── Core ──────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.0.0
anthropic>=0.42.0
openai>=1.60.0
claude-agent-sdk>=0.1.0
Pillow>=11.0.0
python-multipart>=0.0.18
jinja2>=3.1.3
httpx>=0.27.0

#── PII Detection (spaCy optional but recommended) ───────────────────────
#Install with:  pip install spacy && python -m spacy download en_core_web_sm
spacy>=3.7.0

#── Deepfake Detection Agent ─────────────────────────────────────────────
insightface>=0.7.3
onnxruntime>=1.17.0
numpy>=1.26.0
opencv-python>=4.9.0
scipy>=1.12.0
sse-starlette>=2.1.0
aiofiles>=24.1.0
playwright>=1.48.0
